pyproject.toml
[project]
name = "ams"
version = "0.1.0"
description = "A simple ams project"
authors = [
    {name = "vedghimi2054",email = "ghimires2054@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "asgiref (==3.8.1)",
    "diff-match-patch (==20241021)",
    "dj-database-url (==2.3.0)",
    "django (==5.1.6)",
    "django-configurations (==2.5.1)",
    "django-debug-toolbar (==5.0.1)",
    "django-import-export (==4.3.7)",
    "djangorestframework (==3.15.2)",
    "djangorestframework-simplejwt (==5.5.0)",
    "numpy (==2.2.3)",
    "pandas (==2.2.3)",
    "psycopg2 (==2.9.10)",
    "psycopg2-binary (==2.9.10)",
    "pyjwt (==2.9.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.0.1)",
    "pytz (==2025.1)",
    "six (==1.17.0)",
    "smart-getenv (==1.1.0)",
    "sqlparse (==0.5.3)",
    "tablib (==3.8.0)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2025.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Define the packages section
[tool.poetry]
packages = [{include = "ams"}]
package-mode = false  # Disable packaging mode

